requirements.txt
matplotlib==3.5.1
netron==7.1.5
numba==0.48.0
numpy==1.20.0
nuscenes_devkit==1.1.9
opencv_python==4.6.0.66
Pillow==9.5.0
#pycocotools==2.0.4
pycuda==2022.1
pyquaternion==0.9.9
scipy==1.7.3
setuptools==59.8.0
Shapely==1.8.4
terminaltables==3.1.10
thop==0.1.1.post2209072238
tqdm==4.64.0
trimesh==2.35.39
nvidia-pyindex
pytorch-quantization==2.1.3
--extra-index-url https://pypi.ngc.nvidia.com
--extra-index-url https://pypi.tuna.tsinghua.edu.cn/simple
--extra-index-url https://pypi.nvidia.com
